requirements.txt
streamlit
pandas
openai
yagmail
schedule
python-dotenv